requirements.txt
streamlit
pandas
requests
elevenlabs
moviepy
google-generativeai